Adjust structure and remove old config files
