+++ data/config-schema/src/main/java/net/corda/schema/configuration/provider/package-info.java
@@ -1,0 +1,4 @@
+@Export
+package net.corda.schema.configuration.provider;
+
+import org.osgi.annotation.bundle.Export;
