--- data/config-schema/src/main/kotlin/net/corda/schema/configuration/ConfigSchemaException.kt
@@ -1,6 +1,0 @@
-package net.corda.schema.configuration
-
-/**
- * Exception thrown when requested config schema is not available.
- */
-class ConfigSchemaException(msg: String) : Exception(msg)
