--- data/config-schema/src/main/kotlin/net/corda/schema/configuration/SchemaProviderFactory.kt
@@ -1,8 +1,0 @@
-package net.corda.schema.configuration
-
-class SchemaProviderFactory {
-
-    fun getSchemaProvider() : SchemaProvider {
-        return SchemaProviderImpl()
-    }
-}
