+++ data/config-schema/src/main/kotlin/net/corda/schema/configuration/provider/ConfigSchemaException.kt
@@ -1,0 +1,6 @@
+package net.corda.schema.configuration.provider
+
+/**
+ * Exception thrown when requested config schema is not available.
+ */
+class ConfigSchemaException(msg: String) : RuntimeException(msg)
