+++ data/config-schema/src/main/kotlin/net/corda/schema/configuration/provider/SchemaProviderFactory.kt
@@ -1,0 +1,10 @@
+package net.corda.schema.configuration.provider
+
+import net.corda.schema.configuration.provider.impl.SchemaProviderImpl
+
+object SchemaProviderFactory {
+
+    fun getSchemaProvider() : SchemaProvider {
+        return SchemaProviderImpl()
+    }
+}
