--- data/config-schema/src/main/resources/net/corda/schema/configuration/corda.boot.conf
@@ -1,37 +1,0 @@
-# Configuration schema under corda.boot key.
-
-# Messaging configuration to connect to the bus and retrieve the rest of the config. Note that this section
-# mirrors the configuration under the corda.messaging key. Subsequent connections to the bus use a merger of this
-# config and the main corda.messaging config.
-messaging {
-    bus {
-        # Could be DATABASE, KAFKA or INMEMORY
-        type = "KAFKA"
-        # Reflects the corresponding block in the corda.messaging section. When both are available, the corda.messaging
-        # section takes precedence. Note when connecting to the bus to retrieve the configuration, only the boot config
-        # is available. This must therefore be complete enough to establish a connection to the bus.
-        properties {
-            # Settings that apply to both consumers and producers. These are likely to be connection settings which
-            # are similar in both cases.
-            common {
-                # Use standard Kafka configuration properties.
-                bootstrap.servers = "localhost:9092"
-            }
-        }
-        # properties {
-            # When DATABASE
-            # jdbcUrl = "<>"
-            # username = "admin"
-            # passwordSecret = "..."
-        # }
-    }
-}
-
-# The instance ID of this particular worker. Should be unique across instances of the same worker type.
-instanceID = 1
-# Topic configuration. Primarily used to ensure that this worker connects to the correct set of topics on the message
-# bus broker.
-topic {
-    # Global prefix to apply to all topics when subscribing to the bus.
-    prefix = ""
-}
