--- data/config-schema/src/main/resources/net/corda/schema/configuration/corda.crypto.conf
@@ -1,41 +1,0 @@
-corda {
-    cryptoLibrary {
-        // Default Crypto Service Cache parameters
-        keyCache {
-        	// Note: that there will be an instance of the cache object on per Default Crypto Service
-        	// effectively meaning per category + member, so the maximum size could be fairly small
-            persistenceConfig {
-        		// [TBD] - the required parameters to use Kafka based persistence, will be used when the
-        		// the configured crypto service is default
-            }
-      	}
-        //
-        mngCache {
-        	// Note: that there will be an instance of the cache object on per category + member
-        	// so the maximum size could be fairly small
-            persistenceConfig {
-        		// [TBD] - the required parameters to use Kafka based persistence
-            }
-      	}
-        rpc {
-        }
-        cipherSuite {
-        }
-        members {
-            "123" {
-                default {
-                    serviceConfig {
-                    }
-                }
-                "FRESH_KEYS" {
-                    serviceConfig {
-                    }
-                }
-                "LEDGER" {
-                    serviceConfig {
-                    }
-                }
-            }
-        }
-    }
-}
