--- data/config-schema/src/main/resources/net/corda/schema/configuration/corda.db.conf
@@ -1,5 +1,0 @@
-corda {
-    # Only present in the DB Worker config
-    db {
-    }
-}
