--- data/config-schema/src/main/resources/net/corda/schema/configuration/corda.flow.conf
@@ -1,5 +1,0 @@
-corda {
-    flow {
-        # Flow specific overrides here?
-    }
-}
