--- data/config-schema/src/main/resources/net/corda/schema/configuration/corda.messaging.conf
@@ -1,77 +1,0 @@
-# Configuration schema underneath the corda.messaging key. Messaging configuration applies globally across all instances
-# of interaction with the underlying message bus. Configuration items relating to the bus location, timeout settings,
-# batch settings, retries etc can be found in this section. Configuration items applying to individual interactions
-# (e.g. how flow services interact with the bus) are found elsewhere.
-#
-# Note that the boot configuration messaging reflects some of this configuration to ensure that configuration updates
-# can be received in the first place.
-
-# Settings for the underlying bus implementation. These are required to make a connection to the underlying bus
-# implementation.
-bus {
-    type = "KAFKA"
-    properties {
-        # Settings that apply to both consumers and producers. These are likely to be connection settings which
-        # are similar in both cases.
-        common {
-            # Use standard Kafka configuration properties.
-            bootstrap.servers = "localhost:9092"
-        }
-    }
-    # properties {
-        # jdbcUrl = "<>"
-        # username = "admin"
-        # passwordSecret = "..."
-    # }
-}
-# Settings for particular subscription types that read messages from the underlying message bus.
-subscription {
-    # Common settings for all subscriptions. These can be overridden for individual subscription types if required.
-    common {
-        # Timeout for closing the underlying consumers and producers for this subscription. Time is in ms.
-        close.timeout = 400
-        # Time to wait for a response when calling poll on the underlying consumer(s) of a subscription. Time is in ms.
-        poll.timeout = 500
-        # Time to wait when a subscription stops for the processing thread to terminate cleanly. Time is in ms.
-        thread.stop.timeout = 10000
-        # The number of times to retry internal intermittent failures during event processing before throwing to
-        # the client.
-        processor.retries = 3
-        # The number of times to attempt to subscribe to the underlying topic before throwing to the client.
-        subscribe.retries = 3
-        # The number of times to attempt to perform a commit of a transactional process before throwing to the
-        # client.
-        commit.retries = 3
-    }
-    # Settings to apply to all pubsub subscriptions.
-    pubsub = ${subscription.common} {
-    }
-    # Settings to apply to all compacted subscriptions.
-    compacted = ${subscription.common} {
-    }
-    # Settings to apply to all durable subscriptions.
-    durable = ${subscription.common} {
-    }
-    # Settings to apply to all state and event subscriptions.
-    stateAndEvent = ${subscription.common} {
-        # Time to wait for an event processor to succeed. This should be set long enough for events to be processed,
-        # but not so long that the state topic becomes starved due to a lack of polling activity. Time is in ms.
-        processor.timeout = 15000
-    }
-    # Settings to apply to all event log subscriptions.
-    eventLog = ${subscription.common} {
-    }
-    # Settings to apply to all RPC pattern sender-side subscriptions. (Note that this is unrelated to the HTTP RPC
-    # mechanism used to interact with the cluster.)
-    rpcSender = ${subscription.common} {
-    }
-    # Settings to apply to all RPC pattern responder-side subscriptions. (Note that this is unrelated to the HTTP RPC
-    # mechanism used to interact with the cluster.)
-    rpcResponder = ${subscription.common} {
-    }
-}
-# Settings for publishers that write messages to the underlying bus.
-publisher {
-    # Timeout to wait for underlying producers for the publisher. Time is in ms.
-    close.timeout = 600
-}
