--- data/config-schema/src/main/resources/net/corda/schema/configuration/corda.p2p.conf
@@ -1,4 +1,0 @@
-corda {
-    p2p {
-    }
-}
