--- data/config-schema/src/main/resources/net/corda/schema/configuration/corda.platform.conf
@@ -1,5 +1,0 @@
-corda {
-    # platform parameters - version, etc.
-    platform {
-    }
-}
