--- data/config-schema/src/main/resources/net/corda/schema/configuration/corda.policy.conf
@@ -1,4 +1,0 @@
-corda {
-    policy {
-    }
-}
