--- data/config-schema/src/main/resources/net/corda/schema/configuration/corda.rpc.conf
@@ -1,4 +1,0 @@
-corda {
-    rpc {
-    }
-}
